+++ docs/app/[lng]/(main)/layout.tsx
@@ -1,7 +1,32 @@
 import { Container } from '@mui/material'
 
 // Types
+import type { Metadata } from 'next'
 import type { LayoutProps } from '@/types/app/layouts'
+import type { PageProps } from '@/types/app/pages'
+
+// Configurations
+import LocaleConfig from '@/config/locale'
+
+// Helpers
+import { useTranslation } from '@/helpers/i18n/server'
+
+// Metadata
+export const generateMetadata = async (props: PageProps): Promise<Metadata> => {
+	// Props
+	const { params } = props
+
+	// Variables
+	const lng = params?.lng ?? LocaleConfig.default
+
+	// Variables
+	// eslint-disable-next-line react-hooks/rules-of-hooks
+	const { t } = await useTranslation(lng)
+
+	const metadata: Metadata = { title: t('links:home') }
+
+	return metadata
+}
 
 const MainLayout = (props: LayoutProps) => {
 	// Props