+++ docs/app/[lng]/(main)/page.tsx
@@ -7,7 +7,7 @@ import type { PageProps } from '@/types/app/pages'
 import LocaleConfig from '@/config/locale'
 
 // Components
-const IndexTemplate = dynamic(() => import('@/components/templates'))
+const IndexTemplate = dynamic(() => import('@/components/templates/index'))
 
 const HomePage = (props: PageProps) => {
 	// Props