+++ docs/app/mdx-components.tsx
@@ -1,6 +1,0 @@
-// Types
-import type { MDXComponents } from 'mdx/types'
-
-export function useMDXComponents(components: MDXComponents): MDXComponents {
-	return { ...components }
-}