+++ docs/middleware.ts
@@ -11,7 +11,7 @@ import LocaleConfig from '@/config/locale'
 acceptLanguage.languages(LocaleConfig.list)
 
 export const config = {
-	matcher: ['/((?!api|_next/static|_next/image|assets|favicon.ico|sw.js|site.webmanifest|manifest.json|images/).*)']
+	matcher: ['/((?!api|_next/static|_next/image|assets|favicon.ico|sw.js|site.webmanifest|manifest.json|images/|_vercel/|icon.png).*)']
 }
 
 export function middleware(request: NextRequest) {
@@ -19,28 +19,27 @@ export function middleware(request: NextRequest) {
 	const pathname = request.nextUrl.pathname
 	let defaultLng: string | null = null
 
-	if (!defaultLng) defaultLng = acceptLanguage.get(request.headers.get('Accept-Language'))
+	// Response
+	const response = NextResponse.next()
+
 	if (!defaultLng && request.cookies.has(LocaleConfig.cookieName)) defaultLng = acceptLanguage.get(request.cookies.get(LocaleConfig.cookieName)?.value)
+	if (!defaultLng) defaultLng = acceptLanguage.get(request.headers.get('Accept-Language'))
 	if (!defaultLng) defaultLng = LocaleConfig.fallbackLng
 
 	// Redirect if lng in path is not supported
-	if (!LocaleConfig.list.some(lng => request.nextUrl.pathname.startsWith(`/${lng}`)) && !request.nextUrl.pathname.startsWith('/_next'))
-		return NextResponse.redirect(new URL(`/${defaultLng}${request.nextUrl.pathname}`, request.url))
+	if (!LocaleConfig.list.some(lng => pathname.startsWith(`/${lng}`)) && !pathname.startsWith('/_next'))
+		return NextResponse.redirect(new URL(`/${defaultLng}${pathname}`, request.url))
 
 	if (request.headers.has('referer')) {
 		const refererUrl = new URL(request.headers.get('referer') ?? '')
 		const lngInReferer = LocaleConfig.list.find(lng => refererUrl.pathname.startsWith(`/${lng}`))
-		const response = NextResponse.next()
 
 		if (lngInReferer) {
 			response.cookies.set(LocaleConfig.cookieName, lngInReferer)
-			defaultLng = lngInReferer
 		}
 
 		return response
 	}
-	// Find language
-	else defaultLng = defaultLng ?? LocaleConfig.list.find(lng => pathname.startsWith(`/${lng}`))
 
-	return NextResponse.next()
+	return response
 }