+++ docs/types/components/atoms/buttons/text.d.ts
@@ -1,0 +1,12 @@
+// Types
+import type { ReactNode } from 'react'
+
+export type TextButtonProps = {
+	lng?: string
+	title: string
+	startIcon?: ReactNode
+	endIcon?: ReactNode
+	color?: string
+	bordercolor?: string
+	onClick?: () => void
+}