+++ docs/types/components/atoms/typography.d.ts
@@ -1,0 +1,4 @@
+export type TypographyProps = {
+	lng?: string
+	text: string
+}